cleanup(api): remove dead REST fallback branches (retired in M12 Phase D-5)
PR #118 cleaned `src/api/sessions.ts` but left the matching dead
forks in `src/api/content.ts` and the `aux-rest-to-ws.test.ts`
suite that exercised the now-404 REST legs.

Removes:
- 3 `if (shouldUseWs())` forks in `src/api/content.ts`
  (fetchContent, deleteContent, bulkDeleteContent)
- `src/api/aux-rest-to-ws.test.ts` (975 lines testing flag-OFF REST
  branches that no longer compile post-#118)
- Unused `isAuxRestToWsV1Enabled` and `request` imports in content.ts
- Stale REST-fallback commentary

Also fixes the two thread-store tests that broke when #118 dropped
the REST fallback in `getMessagesPage` — they now stub the WS bridge
instead of `globalThis.fetch`.
